+++ Exposure/Player/ExposureFairplayRequester.swift
@@ -45,6 +45,8 @@ extension ExposureFairplayRequester {
         }
         
         print(resourceLoadingRequest.request.url)
+        print(resourceLoadingRequest.redirect?.url)
+        print(resourceLoadingRequest.dataRequest)
         print(resourceLoadingRequest.contentInformationRequest)
         //EMPFairplayRequester only should handle FPS Content Key requests.
         if url.scheme != customScheme {